Casting getpid() return type to something accepted by printf().
darcs-hash:20051118141902-65a35-129e2b0f572ba89a8699a5dd30de15fddb199e54.gz

+++ libsofia-sip-ua/nua/test_nua.c
@@ -4415,7 +4415,7 @@ int main(int argc, char *argv[])
 
   if (o_attach) {
     char line[10];
-    printf("%s: pid %u\n", name, getpid());
+    printf("%s: pid %lu\n", name, (unsigned long)getpid());
     printf("<Press RETURN to continue>\n");
     fgets(line, sizeof line, stdin);
   }
